% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_dates.R
\name{add_key_dates_categorisations}
\alias{add_key_dates_categorisations}
\title{Add columns to cost model categorising line items by the additional columns in 'key dates'}
\usage{
add_key_dates_categorisations(cost_model)
}
\description{
Add columns to cost model categorising line items by the additional columns in 'key dates'
}
