% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneoff_cost.R
\name{add_oneoff_costs}
\alias{add_oneoff_costs}
\title{Add assumptions about oneoff costs to the cost model}
\usage{
add_oneoff_costs(cost_model, oneoff_cost_assumptions)
}
\description{
Add assumptions about oneoff costs to the cost model
}
