% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurring_cost.R
\name{add_recurring_cost}
\alias{add_recurring_cost}
\title{Add assumptions about recurring costs to the cost model}
\usage{
add_recurring_cost(cost_model, recurring_cost_assumptions)
}
\description{
Add assumptions about recurring costs to the cost model
}
