% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff_utilisation.R
\name{add_staff_utilisation}
\alias{add_staff_utilisation}
\title{Add assumptions about staff utilistion to the cost model}
\usage{
add_staff_utilisation(cost_model, staff_utilisation, rate_card)
}
\description{
Add assumptions about staff utilistion to the cost model
}
