% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{add_to_categorisation_columns}
\alias{add_to_categorisation_columns}
\title{Add to the current categorisation columns.  These enable the user to break down the results of the cost model into different categories
such as 'staff costs' and 'hardware costs' at different levels of granularity.}
\usage{
add_to_categorisation_columns(cost_model, categorisation_columns)
}
\description{
Add to the current categorisation columns.  These enable the user to break down the results of the cost model into different categories
such as 'staff costs' and 'hardware costs' at different levels of granularity.
}
