% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_variable_costs.R
\name{add_user_variable_costs}
\alias{add_user_variable_costs}
\title{Add assumptions about recurring costs to the cost model}
\usage{
add_user_variable_costs(cost_model, num_users, user_variable_cost_assumptions)
}
\description{
Add assumptions about recurring costs to the cost model
}
