% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_absolute_increase_to_df_col}
\alias{apply_absolute_increase_to_df_col}
\title{Take a df with a date col, and apply a constant rate of increase to a column}
\usage{
apply_absolute_increase_to_df_col(df, annual_increase, start_date,
  col_to_increase, date_col = "date")
}
\arguments{
\item{annual_increase}{annual increase}
}
\description{
Take a df with a date col, and apply a constant rate of increase to a column
}
