% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_percentage_growth_multiplier_to_df_col}
\alias{apply_percentage_growth_multiplier_to_df_col}
\title{Take a df with a date col, and generate a multiplier corresponding to growth or reduction at a constant percent}
\usage{
apply_percentage_growth_multiplier_to_df_col(df, annual_growth, start_date,
  col_to_increase, date_col = "date")
}
\arguments{
\item{annual_growth}{annual growth rate, where 0.1 = 10\%, -0.05 = -5\% etc.}
}
\description{
Take a df with a date col, and generate a multiplier corresponding to growth or reduction at a constant percent
}
