% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_compatability.R
\name{check_date_compatibility}
\alias{check_date_compatibility}
\title{Check that two dataframes span the same range of dates}
\usage{
check_date_compatibility(this_df, key_dates, this_df_date_col = "date",
  key_dates_date_col = "date")
}
\description{
Check that two dataframes span the same range of dates
}
