% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interpolation.R
\name{constant_growth_interpolation}
\alias{constant_growth_interpolation}
\title{Take a vector and interpolate over the NAs with constant growth}
\usage{
constant_growth_interpolation(vector_with_nas)
}
\description{
Take a vector and interpolate over the NAs with constant growth
}
