% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_excel_dates_in_df}
\alias{convert_excel_dates_in_df}
\title{Take a df with date information, and if there is data information in the format 30/01/17 or 30/01/2017 (Excel outputs this by default) convert to date and warn user}
\usage{
convert_excel_dates_in_df(df, cols = "date")
}
\arguments{
\item{cols}{is either a vector of column names, or NULL. If it's NULL, all columns will be scanned}
}
\description{
Take a df with date information, and if there is data information in the format 30/01/17 or 30/01/2017 (Excel outputs this by default) convert to date and warn user
}
