% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_model.R
\name{create_cost_model}
\alias{create_cost_model}
\title{Create the cost model, a list that stores the various elements of the model}
\usage{
create_cost_model(key_dates)
}
\description{
Create the cost model, a list that stores the various elements of the model
}
