% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_addition_absolute_increase}
\alias{date_to_addition_absolute_increase}
\title{Convert a date into an increase in cost, where costs are growing or shrinking by an absolute amount}
\usage{
date_to_addition_absolute_increase(this_date, start_date, annual_increase)
}
\arguments{
\item{annual_increase}{absolute annual increase}
}
\description{
Convert a date into an increase in cost, where costs are growing or shrinking by an absolute amount
}
