% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_multiplier_percentage_growth}
\alias{date_to_multiplier_percentage_growth}
\title{Convert a date into a multiplier for costs, where costs are growing or shrinking}
\usage{
date_to_multiplier_percentage_growth(this_date, start_date, annual_growth)
}
\arguments{
\item{this_date}{The date to compute a multiplier for.}

\item{start_date}{If \code{this_date = start_date}, the multiplier will be 1}

\item{annual_growth}{annual growth rate, where 0.1 = 10\%, -0.05 = -5\% etc.}
}
\description{
Convert a date into a multiplier for costs, where costs are growing or shrinking
}
