% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_card.R
\name{expand_staff_utilisation_to_time_horizon}
\alias{expand_staff_utilisation_to_time_horizon}
\title{If staff utilisation is provided for a time horizon shorter than key dates, fill in}
\usage{
expand_staff_utilisation_to_time_horizon(staff_utilisation, key_dates,
  frequency)
}
\description{
Zero out all dates outside the time horizon provided
}
