% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interpolation.R
\name{get_all_dates}
\alias{get_all_dates}
\title{Take a dataframe with a date column, and return an ordered list of each day in the date range}
\usage{
get_all_dates(df, date_col = "date", by = "day")
}
\description{
Take a dataframe with a date column, and return an ordered list of each day in the date range
}
