% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_explorer.r
\name{get_cumulative_costs}
\alias{get_cumulative_costs}
\title{Derives cumulative costs, for plotting}
\usage{
get_cumulative_costs(cost_model, cost_dataframe, groupby_vars)
}
\description{
Derives cumulative costs, for plotting
}
