% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_compatability.R
\name{get_min_max_dates}
\alias{get_min_max_dates}
\title{Get the minimum and maximum dates from two dataframes for later compairson}
\usage{
get_min_max_dates(this_df, key_dates, this_df_date_col = "date",
  key_dates_date_col = "date")
}
\description{
Get the minimum and maximum dates from two dataframes for later compairson
}
