% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneoff_cost.R
\name{get_oneoff_cost_chunk}
\alias{get_oneoff_cost_chunk}
\title{Model oneoff costs from an assumptions list}
\usage{
get_oneoff_cost_chunk(assumption_list, key_dates)
}
\description{
Model oneoff costs from an assumptions list
}
