% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneoff_cost.R
\name{get_oneoff_cost_id}
\alias{get_oneoff_cost_id}
\title{Get oneoff costs id
For other modules, each input assumption row can cause many output rows
We therefore split out information about the categorisation of each id, and the rows themselves, to save memory
This function gets all of the relevant categorisation information about an id}
\usage{
get_oneoff_cost_id(assumption_row, cost_model)
}
\description{
Get oneoff costs id
For other modules, each input assumption row can cause many output rows
We therefore split out information about the categorisation of each id, and the rows themselves, to save memory
This function gets all of the relevant categorisation information about an id
}
