% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurring_cost.R
\name{get_recurring_cost_chunk}
\alias{get_recurring_cost_chunk}
\title{Model recurring costs from an assumptions list}
\usage{
get_recurring_cost_chunk(assumption_list, key_dates)
}
\arguments{
\item{key_dates}{a key dates dataframe so the recurring costs have an end date}

\item{assumptions_list}{some assumptions}
}
\description{
Model recurring costs from an assumptions list
}
