% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_variable_costs.R
\name{get_user_variable_costs_chunk}
\alias{get_user_variable_costs_chunk}
\title{Take an assumptions list, a table of the number of users, and key dates, and create a chunk representing user variable costs}
\usage{
get_user_variable_costs_chunk(assumptions_list, users, key_dates)
}
\description{
Take an assumptions list, a table of the number of users, and key dates, and create a chunk representing user variable costs
}
