% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_variable_costs.R
\name{get_user_variable_costs_id}
\alias{get_user_variable_costs_id}
\title{Get a data frame consisting of a single row with the assumption ID and auxiliary information about e.g. the cost category}
\usage{
get_user_variable_costs_id(assumption_list, cost_model)
}
\description{
Get a data frame consisting of a single row with the assumption ID and auxiliary information about e.g. the cost category
}
