% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_xr}
\alias{get_xr}
\title{Get the current exchange rate}
\usage{
get_xr(from, to = "GBP")
}
\description{
This is memoised because otherwise we have to make frequent API calls which slows everything down.
}
