% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interpolation.R
\name{interpolate_days}
\alias{interpolate_days}
\title{Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs to give daily data}
\usage{
interpolate_days(df, date_col = "date", interpolation_fn = zoo::na.approx,
  col_type = is.numeric)
}
\description{
Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs to give daily data
}
