% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interpolation.R
\name{interpolate_days_character}
\alias{interpolate_days_character}
\title{Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs in character columns}
\usage{
interpolate_days_character(df, date_col = "date",
  interpolation_fn = zoo::na.locf)
}
\description{
Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs in character columns
}
