% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interpolation.R
\name{interpolate_days_numeric}
\alias{interpolate_days_numeric}
\title{Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs in numeric columns}
\usage{
interpolate_days_numeric(df, date_col = "date",
  interpolation_fn = zoo::na.approx)
}
\description{
Take a datafrome created by 'get_all_dates_df' and interpolate over the NAs in numeric columns
}
