% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posixt_cols_to_date}
\alias{posixt_cols_to_date}
\title{Convert all datatime (posixt) columns to be dates in a dataframe}
\usage{
posixt_cols_to_date(df)
}
\description{
This is important because data joins on the datetime column can go wrong as a result of things like BST.
}
