% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_assumptions_table}
\alias{process_assumptions_table}
\title{Process an assumptions table by apply a function}
\usage{
process_assumptions_table(assumptions_table, fn)
}
\description{
Process an assumptions table by apply a function
}
