% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneoff_cost.R
\name{process_oneoff_costs}
\alias{process_oneoff_costs}
\title{This function is called when the cost model is run.}
\usage{
process_oneoff_costs(cost_model, this_module)
}
\description{
This function is called when the cost model is run.
}
