% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_variable_costs.R
\name{process_user_variable_costs}
\alias{process_user_variable_costs}
\title{Given a cost model, create the chunks corresponding to user variable costs}
\usage{
process_user_variable_costs(cost_model, this_module)
}
\description{
Given a cost model, create the chunks corresponding to user variable costs
}
