% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_worksheet_from_file_and_tidy}
\alias{read_worksheet_from_file_and_tidy}
\title{Read a worksheet from an Excel file and convert to a tibble.}
\usage{
read_worksheet_from_file_and_tidy(path, sheetname)
}
\description{
This function also ensures that all relevant columns contains dates rather than datetimes.
}
