% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_named_cols_from_df}
\alias{remove_named_cols_from_df}
\title{Removed named columns from a dataframe}
\usage{
remove_named_cols_from_df(df, drops)
}
\description{
Removed named columns from a dataframe
}
