% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_model.R
\name{run_cost_model}
\alias{run_cost_model}
\title{Run the cost model, by iterating through all the registered modules}
\usage{
run_cost_model(cost_model)
}
\description{
Run the cost model, by iterating through all the registered modules
}
