% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{setting_categorisation_columns}
\alias{setting_categorisation_columns}
\title{Overwrite the default categorisation columns.  These enable the user to break down the results of the cost model into different categories
such as 'staff costs' and 'hardware costs' at different levels of granularity.}
\usage{
setting_categorisation_columns(cost_model, categorisation_columns)
}
\description{
Note that the defaults are 'id', 'category_1', 'cateogory_2', 'category_3'
}
