% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{setting_deflator_base_date}
\alias{setting_deflator_base_date}
\title{Set the base date for the deflator (the date for which the gdp deflator = 100)}
\usage{
setting_deflator_base_date(cost_model, base_date)
}
\description{
Set the base date for the deflator (the date for which the gdp deflator = 100)
}
