% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_bubble.R
\name{shiny_bubble}
\alias{shiny_bubble}
\title{A Shiny app that enables the user to explore the cost model}
\usage{
shiny_bubble(cost_model)
}
\description{
A Shiny app that enables the user to explore the cost model
}
