% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_explorer.r
\name{shiny_vis}
\alias{shiny_vis}
\title{A Shiny app that enables the user to explore the cost model}
\usage{
shiny_vis(cost_model)
}
\description{
A Shiny app that enables the user to explore the cost model
}
