% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_duplicated_dates}
\alias{stop_duplicated_dates}
\title{Check that there aren't any duplicated dates}
\usage{
stop_duplicated_dates(df, date_col = "date")
}
\description{
Check that there aren't any duplicated dates
}
