% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_expected_fields}
\alias{stop_expected_fields}
\title{Stop if the list provided does not contains all the expected fields}
\usage{
stop_expected_fields(expected_fields, this_list)
}
\description{
Stop if the list provided does not contains all the expected fields
}
