% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_nonnumeric}
\alias{stop_if_nonnumeric}
\title{Stop data processing if df contains non-numeric data}
\usage{
stop_if_nonnumeric(df, col_names = NULL)
}
\description{
Stop data processing if df contains non-numeric data
}
