% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not_date}
\alias{stop_if_not_date}
\title{Stop data processing if the date column is not of type Date}
\usage{
stop_if_not_date(df, col_name = "date")
}
\description{
Stop data processing if the date column is not of type Date
}
