% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_compatability.R
\name{expand_to_time_horizon}
\alias{expand_to_time_horizon}
\title{Given a dataframe, expand the time horizon out to the minimum and max}
\usage{
expand_to_time_horizon(this_df, key_dates, this_df_date_col = "date",
  key_dates_date_col = "date")
}
\description{
Will throw a warning if this df contains dates outside of the key dates range
}
