% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneoff_cost.R
\name{get_oneoff_cost_chunk}
\alias{get_oneoff_cost_chunk}
\title{Model oneoff costs from an assumptions row, outputting a chunk
Each chunk is a number of rows corresponding to the same input assumption
with expected columns given by the vector}
\usage{
get_oneoff_cost_chunk(assumption_row, key_dates)
}
\description{
Model oneoff costs from an assumptions row, outputting a chunk
Each chunk is a number of rows corresponding to the same input assumption
with expected columns given by the vector
}
