% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurring_cost.R
\name{get_recurring_cost_chunk}
\alias{get_recurring_cost_chunk}
\title{Model recurring costs from an assumptions list}
\usage{
get_recurring_cost_chunk(assumption_list, key_dates)
}
\arguments{
\item{key_dates}{a key dates dataframe so the recurring costs have an end date}

\item{assumptions_list}{this is a list that should have the following elements
\itemize{
\item id: The cost (line item) ID.  This can be used later to join on other information such as the category of spend, etc
\item price_in_original_currency_real: The price of the line item
\item currency: A string.  \code{GBP} if Sterling, otherwise \code{USD}, \code{EUR} etc.
\item frequency:  How often the cost occurs.  \code{day}, \code{week}, \code{month}, \code{year} etc.
\item first date:  a Date. The first time the cost occurs
\item growth_in_real_cost_percent_per_annum
\item growth_in_real_cost_absolute_per_annum
\item growth_in_quantity_percent_per_annum
\item growth_in_quantity_absolute_per_annum
\item quantity
}}
}
\description{
Model recurring costs from an assumptions list
}
